+++ job.c
@@ -144,9 +144,9 @@ rs_result rs_job_iter(rs_job_t *job, rs_buffers_t *buffers)
         if ((orig_in == buffers->avail_in)  &&  (orig_out == buffers->avail_out)
             && orig_in && orig_out) {
             rs_log(RS_LOG_ERR, "internal error: job made no progress "
-                   "[orig_in=%ld, orig_out=%ld, final_in=%ld, final_out=%ld]",
-                   (long) orig_in, (long) orig_out, (long) buffers->avail_in,
-                   (long) buffers->avail_out);
+                   "[orig_in=%f, orig_out=%f, final_in=%f, final_out=%f]",
+                   (double) orig_in, (double) orig_out, (double) buffers->avail_in,
+                   (double) buffers->avail_out);
             return RS_INTERNAL_ERROR;
         }
